use bulk api client directly in bulk action code
